backend/requirements.txt
flask
openai
langchain
langchain-openai
python-dotenv
requests
ragie
langchain-ragie
google-api-python-client
google-auth
tiktoken
langchain-community
qdrant-client
sentence-transformers>=2.2.2
pypdf
python-docx
langchain_cohere
langgraph
langchain-qdrant
pydantic
langchain-google-genai
google-generativeai
Flask-Cors
